requirements.txt
flask
python-dotenv
cryptography
pymongo[srv]
werkzeug
